requirements.txt
amqp==2.6.1
asgiref==3.4.1
billiard==3.6.4.0
celery==4.4.7
celery-progress==0.1.1
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
Django==3.2.9
django-celery-results==2.2.0
django-filter==21.1
django-model-utils==4.2.0
djangorestframework==3.12.4
Faker==9.8.3
importlib-metadata==4.8.2
kombu==4.6.11
Markdown==3.3.6
Pillow==8.4.0
prompt-toolkit==3.0.23
psycopg2==2.9.2
python-dateutil==2.8.0
pytz==2018.9
redis==4.0.2
six==1.12.0
sqlparse==0.4.2
text-unidecode==1.3
vine==1.3.0
wcwidth==0.2.5
wrapt==1.13.3
zipp==3.6.0
